Create and invoke higher order functions
Declare a higher order function called getLength() that takes two
parameters; `arr` and `cb`, it should pass the length of the array
into the callback and invoke it.

Create a higher order function called last()that takes two parameters;
`arr` and `cb`, it should pass the last item of the array into the
callback and invoke it.

Create a higher order function called sumNums() that takes three
parameters; `x`, `y` and `cb`, it should pass the sum of `x` and
`y` into the callback and invoke it.

Create a higher order function called multiplyNums() that takes three
parameters; `x`, `y` and `cb`, it should pass the product of `x` and
`y` into the callback and invoke it.

Create a higher order function called contains() that takes three
parameters; `item`, `list` and `cb`, it should pass `true` or `false`
into the callback based on whether or not `item` is present in `list`
and invoke it.