Perform operations with array methods
Combine the first and last names into a new array called fullName.
Convert each first name into all caps and log the result. Return an
array named largeShirts that contains info about the runners that
have a shirt size of L and log the result. Add up all the donations
into a ticketPriceTotal and log the result.

Extract runners with donations greater than 200 into a goldenDonors
array and log the result. Extract all runners email into an emailList
array and log the result to console. Assign the sum of the golden
donations to a variable called, goldenDonationTotal and log the result
to console.